fix #6207: vm status: return undef values when disk{read, write} cannot be queried
If disk read/write cannot be queried because of QMP timeout, they
should not be reported as 0, because a consumer of the RRD stats
cannot distinguish between 0 being an actual 0 value and 0 being an
indicator for the absence of the real value. The RRD graphs in the UI
will already show this correctly.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250925122829.70121-2-f.ebner@proxmox.com
(cherry picked from commit 528df523161aa424cc6dca69230eae1f9a503cef)
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>