fix #6828: remote migration: bump timeout for writing configuration to accommodate volume activation
The 'config' command will lead to volume activation being done for the
referenced volumes. This is because the 'config' handler in the
mtunnel API endpoint calls into the update_vm_api() function, which
uses the create_disks() function, which is also used for existing
disks. In create_disks(), each volume is activated to do an
existence/basic sanity check by querying its size.

There is no requirement to be fast when handling the 'config' command
during remote migration. Since there could be many disks for a given
VM, allow for up to 2 minutes instead of just 10 seconds.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251001150118.136976-2-f.ebner@proxmox.com