fix #6882: backup provider api: fix backup with TPM state by correctly generating node name
The backup-access API in QEMU expects the '-backup' suffix to be
present for the TPM state fleecing image too. This is a regression of
the switch to using blockdev for fleecing images with commit f92c1fa0
("backup: use blockdev for fleecing images"). Add special handling to
the fleecing_node_name() helper to fix it.

Fleecing backups to native plugins do not use a dedicated image for
the TPM state, so this only affected the backup provider API.

Fixes: f92c1fa0 ("backup: use blockdev for fleecing images")
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251003103319.44974-2-f.ebner@proxmox.com