vm status: make sure disk{read, write} adhere to return schema to fix regression for PDM
As reported in the community forum [0], Proxmox Datacenter Manager
will fail to handle the result of the status API call for a shut down
VM. The reason is that the return schema requires having an integer
return value for disk{read,write}, so an explicit undef is not
allowed. However, the properties are optional and are not relevant for
a shut down guest, so just don't set them at all if no value was
queried. This also makes sure that bug #6207 does not reappear. Note
that pvestatd would already not consider the value for shut down VMs
and treats the value not being present and the value being an explicit
undef the same, see update_qemu_status() in pvestatd.pm.

[0]: https://forum.proxmox.com/threads/160166/post-809901

Fixes: 528df523 ("fix #6207: vm status: return undef values when disk{read, write} cannot be queried")
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251022104259.47945-1-f.ebner@proxmox.com