Add support for Intel TDX
This commit adds support for setting up an Intel TDX VM. A Intel TDX VM
can be setup similar to AMD SEV but uses a different firmware image.

Intel TDX requires kernel_irqchip to be set to `split`, instead of the
default `on`. QEMU does change the default for TDX guests, but it is
included explicitly here for clarity.

Signed-off-by: Philipp Giersfeld <philipp.giersfeld@canarybit.eu>
Signed-off-by: Anton Iacobaeus <anton.iacobaeus@canarybit.eu>
Link: https://lore.proxmox.com/20251028125459.287308-9-anton.iacobaeus@canarybit.eu
[FE: adapt to renamed .fd file]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>