fix #6934: vm start: don't apply pending changes when resuming from hibernation
When loading the VM state during resume from hibernation, it is
essential that the QEMU commandline of the new instance matches the
one from the instance the VM state was created with. Thus, pending
changes cannot be applied.

Checking for hibernation was missing in the logic checking whether
pending changes should be applied. In particular, the $statefile
parameter is not used when resuming after hibernation. There, the
$resume parameter is set and the state file/volume is found in the
'vmstate' configuraiton option.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251029105325.40942-2-f.ebner@proxmox.com
 [TL: resolve trivial context merge conflict]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>