Fix #6989: cloudinit: Enable joliet ISO extension for config disks
The nocloud ISO image disk format mandates 4 files in the root folder
called "network-config", "user-data", "meta-data" and "vendor-data".
These filenames are not ISO9660 Level 3 compatible.

To circumvent this, Proxmox - to date - generates nocloud cloud-init
ISOs with the Rockridge extension (-R) - which adds support for POSIX
compatible filenames and metadata. But the Rockridge extension is not
supported by Windows. In order to also support these filenames for
Windows, also add the joliet extension to generated ISO disks.

Signed-off-by: Markus Ebner <info@ebner-markus.de>
Reviewed-by: Mira Limbeck <m.limbeck@proxmox.com>
Tested-by: Mira Limbeck <m.limbeck@proxmox.com>
Link: https://lore.proxmox.com/mailman.527.1762178392.362.pve-devel@lists.proxmox.com