cpu config: introduce is_abstracted() helper
The helper reports whether the CPU configuration cannot be determined
from the configuration alone. Currently, this means using a custom
model, because that depends on the contents of the custom model
configuration. The next commit will add an abstract flag that depends
on the host.

Note that parse_property_string() returns a hash reference or fails,
so there is no need to check its return value for evaluating to true
in boolean context. Even if the check were necessary, having no result
should be treated as a failure, not quietly ignored.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Reviewed-by: Daniel Kral <d.kral@proxmox.com>
Tested-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20251107144429.126790-4-f.ebner@proxmox.com