clone disk: fix handling of snapshot-as-volume-chain for EFI disks
Currently, cloning an EFI disk off a snapshot-as-volume-chain snapshot
fails with "qemu-img: Failed to load snapshot: Can't find snapshot".
The reason is that the special case for EFI disks calls `qemu-img dd`
with the additional `-l snapname` option, which is only valid for
qcow2-internal snapshots. For snapshot-as-volume-chain snapshots, the
source volume is already the volume corresponding to the snapshot.

Fix this by checking the snapshot method of the volume in question,
and only adding the `-l` option if the volume uses qcow2-internal
snapshots.

Reported in enterprise support.

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20251111123222.73957-1-f.weber@proxmox.com