systemd: vmid.scope: add orderings to fix dbus-vmstate cleanup (regression)
Fixes bug #7092 as well as a regression [0] in combination with HA
shutdown policy 'migrate' caused by commit 1f2c5146 ("migrate: remove
left-over dbus-vmstate instance when migrating without conntrack
state"). Currently, the HA migrations do not migrate the conntrack
state and before this patch, the dbus.service is stopped too early
during shutdown, before the LRM migrations. This led to failure when
trying to check for a left-over dbus-vmstate instance, because
the Net::DBus->system() call would fail.

It is not enough to record the After=dbus.service dependency, because
systemd would already stop the vmid.scope even while processes in it
are still running. Add a Before=pve-ha-lrm.service and for consistency
Before=pve-guests.service to record the ordering on shutdown. This
ensures that the scope units are still around until the LRM or guests
service had the chance to stop, and usually, handle the remaining
running guests itself.

[0]: https://forum.proxmox.com/threads/169739/post-825256

Fixes: https://bugzilla.proxmox.com/show_bug.cgi?id=7092
Co-developed-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Co-developed-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251217131724.118681-3-f.ebner@proxmox.com
Tested-by: Maximiliano Sandoval <m.sandoval@proxmox.com>